fixed no init bug, reorganised template
