--- src/robot-config.cpp
@@ -1,142 +1,0 @@
-#include "lemlib/chassis/trackingWheel.hpp"
-#include "main.h"
-#include "lemlib/api.hpp"
-#include "robot-config.hpp"
-
-// motor groups
-pros::Controller master(pros::E_CONTROLLER_MASTER);
-
-pros::MotorGroup leftMotors({-11, -13, -15}, pros::MotorGearset::blue); // left motor group - ports 3 (reversed), 4, 5 (reversed)
-pros::MotorGroup rightMotors({12, 14, 16}, pros::MotorGearset::blue); // right motor group - ports 6, 7, 9 (reversed)
-
-// ------------ //
-// Other Motors //
-// ------------ //
-
-pros::MotorGroup intake({2, -8}, pros::MotorGearset::blue);
-
-// pros:: Motor smthing else(+-PORT, MotorGearset);
-
-// ---------- //
-// PNEUMATICS //
-// ---------- //
-
-pros::adi::DigitalOut mogo_mech (8);
-
-pros::adi::DigitalOut endgame (7);
-
-// Inertial Sensor on port 10
-pros::Imu imu(1);
-
-// ------------ //
-// ODOM SENSORS //
-// ------------ //
-
-// TODO: init odom sensors
-
-pros::Rotation horizontal_encoder(3);
-
-lemlib::TrackingWheel horizontal_tracking_wheel(&horizontal_encoder, lemlib::Omniwheel::NEW_325, +3.75);
-
-// tracking center:
-    // 8.75, 7.25
-
-// total chassis
-    // 17.5, 14.5
-
-// horizontal wheel
-    // 7.6, 11
-    
-// drivetrain settings
-lemlib::Drivetrain drivetrain(&leftMotors, // left motor group
-                              &rightMotors, // right motor group
-                              15.2, // 11.6 inch track width
-                              lemlib::Omniwheel::NEW_325, // using new 4" omnis
-                              360, // drivetrain rpm is 200 (green direct)
-                              5 // horizontal drift is 2. If we had traction wheels, it would have been 8
-);
-
-// TODO: DrivePID NEEDS TUING
-// lateral motion controller
-lemlib::ControllerSettings linearController(20, // proportional gain (kP)
-                                              0, // integral gain (kI)
-                                              7, // derivative gain (kD)
-                                              0, // anti windup
-                                              0, // small error range, in inches
-                                              0, // small error range timeout, in milliseconds
-                                              0, // large error range, in inches
-                                              0, // large error range timeout, in milliseconds
-                                              0 // maximum acceleration (slew)
-);
-// angular motion controller
-
-// TODO: TUNE KI, Others should be fine, KI might be too big now but the previous value of 0.01 was too small
-
-lemlib::ControllerSettings angularController(4.0, // proportional gain (kP)
-                                             0.30, // integral gain (kI)
-                                             21.3, // derivative gain (kD)
-                                             7, // anti windup
-                                             0.5, // small error range, in degrees
-                                             1000, // small error range timeout, in milliseconds
-                                             2, // large error range, in degrees
-                                             1000, // large error range timeout, in milliseconds
-                                             0 // maximum acceleration (slew)
-); 
-
-// input curve for throttle input during driver control
-lemlib::ExpoDriveCurve throttleCurve(3, // joystick deadband out of 127
-                                     10, // minimum output where drivetrain will move out of 127
-                                     1.019 // expo curve gain
-);
-
-// input curve for steer input during driver control
-lemlib::ExpoDriveCurve steerCurve(3, // joystick deadband out of 127
-                                  10, // minimum output where drivetrain will move out of 127
-                                  1.019 // expo curve gain
-);
-
-// sensors for odometry
-lemlib::OdomSensors sensors(nullptr, // vertical tracking wheel
-                            nullptr, // vertical tracking wheel 2, set to nullptr as we don't have a second one
-                            &horizontal_tracking_wheel, // &horizontal_tracking_wheel
-                            nullptr, // horizontal tracking wheel 2, set to nullptr as we don't have a second one
-                            &imu // inertial sensor &imu
-);
-
-lemlib::Chassis chassis(drivetrain, linearController, angularController, sensors, &throttleCurve, &steerCurve);
-
-void initialize() {
-    pros::lcd::initialize(); // initialize brain screen
-    chassis.calibrate(); // calibrate sensors
-    intake.set_brake_mode(pros::E_MOTOR_BRAKE_HOLD);
-
-
-    pros::Task screenTask([]() {
-        while (true) {
-            pros::lcd::print(0, "X: %f", chassis.getPose().x); // x
-            pros::lcd::print(1, "Y: %f", chassis.getPose().y); // y
-            pros::lcd::print(2, "Theta: %f", chassis.getPose().theta); // heading
-            lemlib::telemetrySink()->info("Chassis pose: {}", chassis.getPose());
-            pros::delay(50);
-        }
-    });
-}
-
-// Implement error hdling later
-
-void auton_init()
-{
-    chassis.setBrakeMode(pros::E_MOTOR_BRAKE_HOLD);
-    // Additional auton init code
-}
-
-
-
-void controller_controls()
-{
-    int leftY = master.get_analog(pros::E_CONTROLLER_ANALOG_LEFT_Y);
-    int rightX = master.get_analog(pros::E_CONTROLLER_ANALOG_RIGHT_X);
-    // move the chassis with curvature drive
-    chassis.curvature(leftY, rightX);
-}
-
